正在完善服务器端软件的业务功能

+++ clustermanage-server/src/main/java/cc/bitky/clustermanage/server/message/web/WebMsgDeployRemainChargeTimes.java
@@ -16,7 +16,7 @@ public class WebMsgDeployRemainChargeTimes extends BaseMessage {
         this.times = times;
     }
 
-    public long getTimes() {
+    public int getTimes() {
         return times;
     }
 }
